+++ tests/Honua.Server.Tests/Features/Admin/StreamingOperationsEndpointsTests.cs
@@ -4,7 +4,7 @@
 using System.Net;
 using System.Text.Json;
 using FluentAssertions;
-using Honua.Server.Features.Alerts;
+using Honua.Server.Features.Infrastructure.Abstractions;
 using Honua.TestKit;
 using Honua.TestKit.Attributes;
 using Honua.TestKit.Constants;
