+++ tests/Honua.Server.Tests/Features/Alerts/AlertsServiceCollectionExtensionsTests.cs
@@ -5,6 +5,7 @@ using FluentAssertions;
 using Honua.Core.Features.Alerts.Abstractions;
 using Honua.Core.Features.Alerts.Domain;
 using Honua.Server.Features.Alerts;
+using Honua.Server.Features.Infrastructure.Abstractions;
 using Honua.TestKit.Attributes;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
